requirements.txt
amqp==1.4.9
androguard==3.0.1
anyjson==0.3.3
args==0.1.0
asn1crypto==0.23.0
billiard==3.3.0.23
celery==3.1.25
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
clint==0.5.1
cryptography==2.0.3
Django==1.11.5
django-celery==3.2.1
django-filter==1.0.4
djangorestframework==3.6.4
enum34==1.1.6
futures==3.1.1
GPlayCli==2.15
idna==2.6
ipaddress==1.0.18
kombu==3.0.37
Logbook==1.1.0
lxml==4.0.0
Markdown==2.6.9
mistune==0.7.4
ndg-httpsclient==0.4.3
protobuf==3.4.0
psycopg2==2.7.3.1
py==1.4.34
pyasn1==0.3.6
pycparser==2.18
pycrypto==2.6.1
pyOpenSSL==17.3.0
pyshark==0.3.7.9
pytz==2017.2
PyYAML==3.12
requests==2.18.4
six==1.11.0
trollius==1.0.4
urllib3==1.22
vine==1.1.4
